+++ src/BEST/Index/single_threaded.hpp
@@ -28,9 +28,9 @@ class SingleThreadedIndex  : public NGramBtreeIndex {
      * R_c: set of idx of r s.t. \exists some g \in r
      */
     struct job {
-        std::vector<std::set<unsigned int>> qg_list;
-        std::vector<std::vector<unsigned int>> gr_list;
-        std::vector<unsigned int> rc;
+        std::vector<std::set<size_t>> qg_list;
+        std::vector<std::vector<size_t>> gr_list;
+        std::vector<size_t> rc;
     };
 
     SingleThreadedIndex() = delete;
@@ -66,12 +66,12 @@ class SingleThreadedIndex  : public NGramBtreeIndex {
 
     std::vector<std::string> candidate_gram_set_gen(
         std::vector<std::vector<std::string>> & query_literals,
-        std::map<std::string, unsigned int> & pre_suf_count);
+        std::map<std::string, size_t> & pre_suf_count);
 
-    std::map<std::string, unsigned int> get_all_gram_counts(
+    std::map<std::string, size_t> get_all_gram_counts(
         const std::vector<std::vector<std::string>> & query_literals);
     
-    std::unordered_map<unsigned int, std::vector<size_t>> k_medians(
+    std::unordered_map<size_t, std::vector<size_t>> k_medians(
         const std::vector<std::vector<double>> & dist_mtx, 
         int num_queries, int num_clusters);
     
@@ -80,32 +80,32 @@ class SingleThreadedIndex  : public NGramBtreeIndex {
     
     std::vector<std::vector<double>> calculate_pairwise_dist(
         const std::vector<std::set<std::string>> & qg_gram_set,
-        const std::map<std::string, unsigned int> & pre_suf_count);
+        const std::map<std::string, size_t> & pre_suf_count);
     
     void compute_benefit(std::vector<long double> & benefit, 
-        const std::set<unsigned int> & index, 
+        const std::set<size_t> & index, 
         const best_index::SingleThreadedIndex::job & job, 
         size_t num_queries);
     
-    bool all_covered(const std::set<unsigned int> & index,
+    bool all_covered(const std::set<size_t> & index,
         const best_index::SingleThreadedIndex::job & job,  
         size_t query_size);
     
     void indexed_grams_in_string(const std::string & l, 
         const std::vector<std::string> & candidates,
-        std::vector<std::set<unsigned int>> & g_list, 
+        std::vector<std::set<size_t>> & g_list, 
         size_t idx,
         const std::vector<bool> & candidates_filter=std::vector<bool>());
 
     void indexed_grams_in_literals(const std::vector<std::string> & literals, 
         const std::vector<std::string> & candidates,
-        std::vector<std::set<unsigned int>> & g_list,
+        std::vector<std::set<size_t>> & g_list,
         size_t idx);
     
     void build_gr_list_rc(best_index::SingleThreadedIndex::job & job, 
         size_t candidates_size,  
-        unsigned int dataset_size,
-        const std::vector<std::set<unsigned int>> & rg_list);
+        size_t dataset_size,
+        const std::vector<std::set<size_t>> & rg_list);
 
  private:
     long double k_reduced_queries_size_;
@@ -115,13 +115,13 @@ class SingleThreadedIndex  : public NGramBtreeIndex {
 
     /** Helpers **/
     void workload_reduction(std::vector<std::vector<std::string>> & query_literals,
-        std::map<std::string, unsigned int> & pre_suf_count);
+        std::map<std::string, size_t> & pre_suf_count);
 
-    bool index_covered(const std::set<unsigned int> & index, 
+    bool index_covered(const std::set<size_t> & index, 
         const best_index::SingleThreadedIndex::job & job,
         size_t r_j, size_t q_k);
 
-    void build_qg_list(std::vector<std::set<unsigned int>> & qg_list,
+    void build_qg_list(std::vector<std::set<size_t>> & qg_list,
         const std::vector<std::string> & candidates, 
         const std::vector<std::vector<std::string>> & query_literals);
